Fix episode deletion
It was not possible to delete an episode if an asset was linked to it
